+++ app/assets/js/src/typeguardian/writeTypeguardExpression.ts
@@ -1,10 +1,20 @@
 import { writeIndentation } from './writeIndentation.js';
-import { writeTypeGuardName } from './writeTypeguardName.js';
+import { writeTypeguardName } from './writeTypeguardName.js';
+
+interface WriteTypeguardExpressionOptions {
+	indent: string;
+	indentLevel: number;
+	// TODO: Add option to use nested assertions
+}
 
 /**
  * Write a typeguard expression for a single type, i.e. not an object with properties.
  */
-export function writeTypeguardExpression(propName: string, propType: string, indent = '    ', indentLevel = 0): string {
+export function writeTypeguardExpression(propName: string, propType: string, options: WriteTypeguardExpressionOptions): string {
+	const {
+		indent,
+		indentLevel,
+	} = options;
 	const baseIndent = writeIndentation(indent, indentLevel);
 
 	/** These primitives can be checked via the `typeof` operator */
@@ -37,8 +47,11 @@ export function writeTypeguardExpression(propName: string, propType: string, ind
 ${baseIndent}${indent}Array.isArray(data.${propName}) &&
 ${baseIndent}${indent}data.${propName}.every(${
 	innerType.match(customTypePattern)
-		? writeTypeGuardName(innerType)
-		: `() => ${writeTypeguardExpression(propName, innerType, indent, indentLevel + 1)}`
+		? writeTypeguardName(innerType)
+		: `() => ${writeTypeguardExpression(propName, innerType, {
+			indent,
+			indentLevel: indentLevel + 1,
+		})}`
 })
 ${baseIndent}`;
 	}
@@ -54,14 +67,17 @@ ${baseIndent}`;
 
 	if (unionMembers.length > 1) {
 		return `
-${baseIndent}${indent}${unionMembers.map((type) => writeTypeguardExpression(propName, type, indent, indentLevel + 1)).join(` ||\n${baseIndent}${indent}`)}
+${baseIndent}${indent}${unionMembers.map((type) => writeTypeguardExpression(propName, type, {
+	indent,
+	indentLevel: indentLevel + 1,
+})).join(` ||\n${baseIndent}${indent}`)}
 ${baseIndent}`;
 	}
 
 	// Custom types
 	const isCustomType = (customTypePattern).test(propType);
 	if (isCustomType) {
-		return `${writeTypeGuardName(propType)}(data.${propName})`;
+		return `${writeTypeguardName(propType)}(data.${propName})`;
 	}
 
 	// Unrecognised pattern, left to developer to implement